LLM Flight Assistant/requirements.txt
python-dotenv>=1.0.0
requests>=2.26.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
nltk>=3.6.0